Renamed PaymentUtils to PaymentHelper and take care of PR comments
